+++ migrations/add_logs_pointeuse.sql
@@ -1,0 +1,4 @@
+ALTER TABLE configlspd ADD COLUMN logs_pointeuse VARCHAR;
+
+-- Insère l'ID du salon pour la colonne logs_pointeuse
+UPDATE configlspd SET logs_pointeuse = '1426730229614903326';